requirements.txt
pexpect==4.7.0
colorlog==4.0.2
python-telegram-bot==12.2.0
PyYAML==5.1.2
pycrypto==2.6.1
